feat: rudimentary upload flow
