+++ backend/requirements.in
@@ -1,3 +1,4 @@
 fastapi~=0.101
 uvicorn[standard]
+python-multipart
 psycopg2
