+++ backend/requirements.txt
@@ -24,6 +24,8 @@ pydantic-core==2.4.0
     # via pydantic
 python-dotenv==1.0.0
     # via uvicorn
+python-multipart==0.0.6
+    # via -r requirements.in
 pyyaml==6.0.1
     # via uvicorn
 sniffio==1.3.0
